IOS-3337 add RavencoinMultiNetworkProvider
